refactor(consensus): Create an IDKG crate (#4526)
This PR factors the `idkg` module from consensus out into it's own
crate.

There is a module called `test_utils` as part of `idkg`, which is also
used for tests in the main `consensus` module.
Since `#[cfg(test)]` flagged modules are not compiled for dependencies,
this is an issue.

This PR solves this by moving some of the
`ic_consensus::idkg::test_utils.rs` code into
`ic_test_utilities_consensus::idkg.rs`. It is not possible to move all
code there, since that would create a circular dependency.

As part of this PR, I only moved the code there, that was necessary to
sepparate `idkg` from `consensus`. In the future, we might to move even
more code from `test_utils` there.

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>