feat(sns): Implement topic following index (#4576)
This PR implements a dynamic index for the topic-based follower index
relation. The index will be used to implement the new, topic-based
following. Note that the neurons themselves store the "follows" relation
which is the inverse of "is-followed-by".

This PR also fixes a small mistake in the new API (switching from
`uint64` to `int32` indices). This is safe because the API corresponds
to a command that has not yet been enabled in production.