refactor: Move system_api inside embedders crate (#4591)
The System API implementation is in a separate crate for historical
reasons. However, a lot of the plumbing that needs to happen when adding
a new API is in the embedders crate (including validation of the Wasm
module, linking the imports to the implementation, etc). This PR moves
the implementation in the embedders crate to make it more contained and
easier for people to make all changes they need for a new API in a
single place.

High-level changes included:
1. Move `system_api` under `embedders/src/system_api` folder mostly
preserving its structure.
2. Previous `system_api/lib.rs` moved to `embedders/src/system_api.rs`
(as the module's entry point).
3. Updated all imports and bazel/cargo files to reflect the new
structure.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>