feat(PocketIC): new function PocketIcBuilder::with_read_only_state_dir (#4495)
This PR adds a new function `PocketIcBuilder::with_read_only_state_dir`
to specify a directory from which the state of the PocketIC instance
should be loaded without modifying the provided directory (i.e., the
provided directory is "read-only"). This is useful, e.g., to resume
multiple instances from the same state concurrently (in which case
concurrent modifications of the same underlying state directory would
lead to data races).