chore(fuzzing): Cleanup AFL flags (#4650)
AFLPlusPlus currently flags any env var with AFL_ prefix as `Mistyped
AFL environment variable:` if the variable is not present in a
[predefined list
](https://github.com/AFLplusplus/AFLplusplus/blob/5a527046a5b0035a556ddeca2c0ad1c7e68e8f03/src/afl-fuzz-state.c#L153).
This PR aims to cleanup such envvar to avoid misleading logs

1. `AFL_DRIVER_DONT_DEFER` is an old flag used to skip afl deferred mode
when using afl_driver.cpp (a glue between AFL++ and
LLVMFuzzerTestOneInput harnesses). We use afl_driver.cpp to use the same
code bindings for both AFL and libfuzzer. Initially, we had some build
issues with this driver and deferred mode but things are stable now and
the flag can be removed.

2. Rename `AFL_RELEASE_VERSION` to `AFLPLUSPLUS_RELEASE_VERSION` as that
is more appropriate

---------

Co-authored-by: IDX GitHub Automation <>
Co-authored-by: Bas van Dijk <bas@dfinity.org>