fix(IDX): Extract CI build config
This extracts the CI build config (whether to diff the targets, to run a
release build, etc) into a new job so that it is centralized and so that
other jobs (`bazel-test-all`, `build-ic`) can depend on it.

This also adds a job summary so that the build config can be checked
easily.