fix(IDX): Extract CI build config (#4662)
This extracts the CI build config (whether to diff the targets, to run a
release build, etc) into a new job so that it is centralized and so that
other jobs (`bazel-test-all`, `build-ic`) can depend on it.

This also adds a job summary so that the build config can be checked
easily.

---------

Co-authored-by: Marko Kosmerl <marko.kosmerl@dfinity.org>
Co-authored-by: IDX GitHub Automation <>