feat(node_reward): Add the basic calculation to the Node Reward Canister (#4628)
# What
We are adding the same calculation to the Node Reward canister that the
Registry canister currently has. In addition, we are exposing the
ability of this canister to specify which version of the data the
rewards are calculated against.

# Why 
This is part of the migration of rewards out of Registry into this new
canister, and will allow back-testing previous values for correctness,
as well as enhancing auditability.