sns-testing: make ICP treasury configurable
Currently, the identity that transfer ICPs to the swap participants or
the user provided account is hardcoded. Which limits the usage of the
'sns-testing' to the network created by 'sns-testing-init'.

This commit adds an optional argument to choose the dfx identity to be
used as an ICP treasury to 'sns-testing-init' binary and
'swap-complete' and 'transfer-icp' subcommands of 'sns-testing'.