chore(NODE-1579): fix flakiness in nested systests by only running them in dm1
The Farm hosts in the dm1 DC run:
Ubuntu-24.04, libvirt-10.0.0 and QEMU-8.2.2 while the other (older) Farm hosts run
Ubuntu-20.04, libvirt-6.6.0  and QEMU-5.0.0.
If the host VM is allocated to these older hosts the nested guest VM often runs into soft CPU lockups.
So we temporarily require that this test is hosted in dm1 until the other Farm hosts are upgraded to Ubuntu 24.04.