feat(nns): Always panic when pb methods are called (#4696)
# Why

The `_pb` method in the NNS Governance are deprecated. This PR makes
them always fail.

# What

Make the failure probability 100% for the `_pb` methods. We could simply
panic. However, that will require us to remove branches of code, which
might make it more difficult to rollback if needed if subsequent code is
relying on the removal.