test: Add checks for Traps in the Random Traffic Canister Framework (#4689)
- Adds a check for traps in the canister logs since there are asserts in
the canister that would fire for misrouted or duplicate responses for
example.
- Improves the test about canister migration to really enforce what it
is supposed to test rather than rely on randomness to occasionally
produce this situation since I was unable to observe that during manual
inspection.
- The default number of instructions was bumped since making this lower
does not speed up the test, but rather it can turn the heartbeat into an
apparent no-op by turning aborting and resetting. This unfortunately
does not show up in the logs; a comment warning about this was added
instead and a counter for successful heartbeat invocations such that a
no-op there can be distinguished from other problems.
- Some refactoring.