feat(nns): Allow Governance to request rewards from Node Reward canister instead of registry (behind flag) (#4693)
# What

This allows a test version of governance to rely on the new Node Rewards
canister to calculate node rewards. In the normal case, it will continue
to rely on Registry.

We also fix an issue in the Node Reward canister, where a query was used
when an update should have been used. Since this canister is not yet
considered "live" and this API endpoint should not yet be relied upon,
we are overriding the DIDC_CHECK on backwards compatibility.

# Why

This is part of the effort to allow node rewards to be calculated based
on useful work.