feat(nns): Define VotingPowerSnapshot and a collection of snapshots (#4404)
# Why

To prevent a sudden takeover of the NNS in the event of some party
gaining illegitimate voting power (e.g. a bug in Governance, ICP Ledger,
etc.), we aim at reverting the voting power to an earlier snapshot, so
that neurons don't gain voting power immediately if a spike is detected.

# Related PRs

There are 3 PRs to achieve the above-mentioned goal:

- (current PR) Define VotingPowerSnapshot and a collection of snapshots
- (#4405) Add a timer task to record snapshots periodically
- (#4406) Use the previous snapshot if a voting power spike is detected

# What

* Define a VotingPowerSnapshot to represent the voting power calculated
at a certain point of time
* One key decision is to also snapshot the voting power of each neuron,
which then can be used to create ballots later, if (in an extremely
unlikely event) a voting power spike is detected
* Implement how the snapshot can be used:
  * Add a snapshot
* Check the last time the snapshot is added (for scheduling tasks to add
snapshots)
  * Check if the voting power has a spike