chore(IDX): use execution log for determinism checks (#4771)
This changes the build determinism checks to use the JSON execution log
produced by Bazel.

This ensures all steps of the builds are checked and provides a
"stacktracke" in case a determism issue occurs (the output will show the
list of dependencies that have changed all the way to the root cause).

This also means we can plug this directly into the bazel build (bazel
action) without having to compare artifacts manually; in particular this
avoids jobs like build-ic & bazel-test-all getting out of sync and
producing different targets.

---------

Co-authored-by: IDX GitHub Automation <>