test(ICRC_Ledger): Remove migration-related checks in ICRC ledger suite golden state test (#4782)
Since the migration to stable structures is complete, it shall again be
possible to roll back to the previous ledger version. This PR proposes
to remove the parameters and functionality that allowed for failed
downgrades to not fail the test - since the downgrades tested here
should now always be possible, any downgrade failure should also cause
the test to fail.