feat(nns): Define NeuronAsyncLock to be compatible with safer access pattern to global state (#4774)
# Why

The `LedgerUpdateLock` contains a `*mut Governance` which requires
`unsafe` in order to unlock, and it is only compatible with the
`governance_mut()` access pattern where `unsafe` is also used for
accessing the canister state. As we are moving towards accessing the
canister state through `LocalKey<RefCell<Governance>>`, a new access
pattern for neuron locks is needed.

# What

* Move `LedgerUpdateLock` and the locking/unlocking mechanisms to
`neuron_lock.rs`
* Implement `NeuronAsyncLock` to be compatible with `&'static
LocalKey<RefCell<Governance>>`
* Implement locking and unlocking mechanisms similar to
`LedgerUpdateLock`