feat(nns): Add a timer task to perform voting power snapshots (#4405)
# Why

To prevent a sudden takeover of the NNS in the event of some party
gaining illegitimate voting power (e.g. a bug in Governance, ICP Ledger,
etc.), we aim at disabling the early adoption of NNS proposals if the
voting power has a spike, which should give the network more time to
react and recover in such cases.

# Related PRs

There are 3 PRs to achieve the above-mentioned goal:

- (#4404) Define VotingPowerSnapshot and a collection of snapshots
- (current PR) Add a timer task to record snapshots periodically
- (#4798) Use the previous snapshot if a voting power spike is detected

# What

* Refactor the existing function `create_ballots_for_standard_proposal`
to be used for snapshots too
* Define how to store the snapshots in stable storage
* Define a timer task for taking snapshots