chore(nns): Delete flags related to periodic confirmation of following. (#3782)
The "Periodic Confirmation of Following/Voting Power" feature started
having its full effect at the beginning of April, and no reason to
recall it has emerged (knock on wood). DFINITY has looked at the
practical consequences of this, and it does not seem like any
interventions are needed. Therefore, now is a good time to get rid of
these flags.

As with all such clean ups, it is not super important that this go in
soon. If there are any objections to merging this now, then we can
definitely delay.

# References

Closes [NNS1-3601].

[NNS1-3601]: https://dfinity.atlassian.net/browse/NNS1-3601