test(IDX): move //rs/execution_environment:execution_environment_misc_integration_tests/dts to its own target (#4711)
The P90 duration of the
[//rs/execution_environment:execution_environment_misc_integration_tests/dts](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/execution_environment/tests/dts.rs)
test [jumped
up](https://superset.idx.dfinity.network/explore/?slice_id=108) starting
in March 2024 and reached above the maximum of 5 minutes. We can't
reproduce this locally so it's likely related to a load issue on CI.

To have more control over the resource usage of this test we move it to
its own dedicated bazel target ` //rs/execution_environment:dts_test`.
We don't change any CPU reservation or `RUST_TEST_THREADS` settings yet
such that we can gather some data on the baseline durations.