fix(ICRC_Ledger): FI-1709: Recompute ICRC ledger certified data in post upgrade (#4796)
Recompute the certified data in the post upgrade of the ICRC ledger.

The certified data is currently only updated during ledger init, and
after a transaction. This means that, without this fix, with the ICRC-3
certificate update that fixes the non-compliant labels, the ledger will
return an invalid certificate until it process its first transaction
after the canister upgrade.