feat: [CON-1469] periodically fetch the nns delegation (#3902)
Currently we do it only at the start of the replica. With the proposed
change we will do it every 10 minutes.

# Tests

Added several unit tests and a system test which reads sends a
`read_state` requests to app and nns subnets and checks that eventually
the returned delegation's timestamp increases. Note that the system test
is quite slow, because it will take at least 10 minutes to finish, as it
waits until the nns delegation is refreshed.

# Metrics
In https://github.com/dfinity/ic/pull/4017 we introduced new metrics
which gave us some insight into how big the delegations are and how long
it takes to fetch them.

## Delegation size

- max => 660 bytes
- median => 637 bytes
- min => 485 bytes

## Delegation fetch time
- max => 5 minutes
- median => 0.6 seconds

Assuming that each delegation has at most 660 bytes and each node on
non-nns subnet (in total around 500) fetches a delegation from a random
nns node every 10 minutes we get:
1) 500 / (40 * 10) = 1.25 requests per nns node per minute
2) which corresponds to roughly 1KB of requests per nns node per minute
which shouldn't introduce too much load on the NNS nodes

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>