sns-testing: make specifying PocketIC state directory optional
Currently, 'sns-testing-init' requires specifying the path to the
PocketIC directory explicitly. This introduces the necessity to clean up
the directory before re-running this binary (if the same value for the
option is used).

This commit makes '--state-dir' argument optional. If it's not provided,
a new temporary directory will be created. This directory will remain
even after 'pocket-ic-server' process has finished. This temporary
directory will be cleaned up on the next system reboot.