chore: [MR-675] Deserialize wasm with hash always present (#4734)
[MR-675](https://dfinity.atlassian.net/browse/MR-675)

`binary_hash` was
[introduced](https://github.com/dfinity/ic/commit/61958e6dfdc9959ea0cb28b6de9dbc91a4434f71)
as an optional field in `ExecutionStateBits` for backwards compatibility
reasons. That was because when loading checkpoints written by even older
replica versions, `binary_hash` would not exist.

In production, we have populated `binary_hash` with `Some` value for
long time and serialize it to disk. It is OK to remove the optional tag
and deserialization won't have issues.

In addition to simplifying code, my primary goal of this PR is to get
rid of the usage of `canister_module.as_slice()` when deserializing Wasm
files so that we know for sure that it will be always lazy loading after
we implement [MR-651](https://dfinity.atlassian.net/browse/MR-651) in
the future.

The `binary_hash`  in `CanisterSnapshotBits` follows the same pattern.


[MR-675]:
https://dfinity.atlassian.net/browse/MR-675?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[MR-651]:
https://dfinity.atlassian.net/browse/MR-651?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ