test(consensus): Increase number of retries of get signature in system tests (#4835)
This PR increases the number of retries when trying to acquire a chain
key signature in the system tests.

Certain tests activate the chain key signatures just before trying to
get a signature. At that time, the routing of the IC might not be ready
yet. Currently, we retry only 5 times, with a 2 second backoff. This can
lead to flackyness.