chore: CON-1478 Grant backup user read access to artifact pools (#4817)
In case of an NNS recovery, we may want to recover using backup data. In
order to make sure we recover with the correct latest state, we need to
compare the local state with the latest certifications on the subnet.
However, certifications are currently not part of the backup data.

This PR extends the read-only access of the backup user to the artifact
pools. This allows backup pod operators to create backups of
certification artifacts on demand during a potential recovery.

The backup access is extended by introducing a new group
`ic-consensus-pool` and adding the current GuestOS users, including
`backup`, similar to how it was done previously for the
[ic-registry-local-store](https://github.com/dfinity/ic/commit/39ced3704fcfae47187b2b3e22699777d69843ca).