refactor(ic-canisters-http-types): Copy crate ic-canisters-http-types to /packages in preparation for publication (#4849)
Canisters such as
[sol-rpc-canister](https://github.com/dfinity/sol-rpc-canister/) and
[evm-rpc-canister](https://github.com/dfinity/evm-rpc-canister) need to
implement the HTTP gateway protocol to be able to deliver logs and
metrics via HTTPs. Currently we keep copying around `HttpRequest` and
`HttpResponse`defined in the
[ic-canister-http-types](https://github.com/dfinity/ic/tree/master/rs/rust_canisters/http_types)
crate.

The goal of this PR is to copy the
[ic-canister-http-types](https://github.com/dfinity/ic/tree/master/rs/rust_canisters/http_types)
crate to packages in preparation for publication on
[crates.io](crates.io). This PR also adds extra documentation and unit
tests to the original
[ic-canister-http-types](https://github.com/dfinity/ic/tree/master/rs/rust_canisters/http_types)
crate.

A second PR will be opened to clean the internal
[ic-canister-http-types](https://github.com/dfinity/ic/tree/master/rs/rust_canisters/http_types)
crate and make sure other dependancies in the repo use the newly
published
[ic-http-types](https://github.com/dfinity/ic/tree/mducroux/XC-313-ic-canisters-http-types-crate/packages/ic-http-types)
crate. (cc @mraszyk)