test(sns): Porting sns-testing to the ICP mono repo, Part III (#4705)
This is the third PR for implementing the next-gen
[sns-testing](https://github.com/dfinity/sns-testing) tool.

At the core of this tool lies the [PocketIC
framework](https://crates.io/crates/pocket-ic/6.0.0), the testing
framework capable of creating a fully deterministic environment for
testing complex ICP dapp scenarios.

This PR is authored by Serokell and reviewed by the DFINITY Foundation.

< [Previous PR](https://github.com/dfinity/ic/pull/4548) |

---------

Co-authored-by: Roman Melnikov <roman.melnikov@serokell.io>
Co-authored-by: Roman Melnikov <rvembox@gmail.com>