refactor: Move unflushed_changes out of snapshots
CanisterSnapshots contains a field unflushed_changes that communicates
to the checkpoint logic what snapshot creations/deletions/restores
happened so that they can be correctly represented on disk.

The same field also needs to be used for some of the upcoming canister
migration logic, in particular when a canister is renamed locally.

It would be weird for the canister migration logic to populate a field in
CanisterSnapshots, so this refactor moves the field out to a more
general location. There should be no change in behaviour.