fix: update parsing of power measurements via IPMI tool (#4833)
Deployed to bare-metal test machine and scraped the metrics:
```
# HELP power_average_watts Average power reading, Watts
# TYPE power_average_watts gauge
power_average_watts 143
# HELP power_instantaneous_watts Instantaneous power reading, Watts
# TYPE power_instantaneous_watts gauge
power_instantaneous_watts 143
# HELP power_sampling_period_seconds Power sampling period, seconds
# TYPE power_sampling_period_seconds gauge
power_sampling_period_seconds 1
```
The issue was that when the timer runs, the output of the IPMI tool is
slightly different from when it is run manually. Manually, it includes
"AM"/"PM" with the time. Hence, the offset by 1 in the `awk` expression.