feat(registry): Added `maybe_chunkify`... (#4751)
... and have it be called from `changelog_insert`.

The new behavior is behind a flag, and as per usual, in release builds,
it is disabled.

`maybe_chunkify` takes a `RegistryAtomicMutateRequest`, and converts it
to an equivalent `HighCapacityRegistryAtomicMutateRequest`.

When the input is "too large" (and this feature is enabled), instead of
having the value inlined, it is stored in stable memory, and the inline
value is replaced with a "pointer" to the stable memory.

# References

[NNS1-3681]

[NNS1-3681]: https://dfinity.atlassian.net/browse/NNS1-3681

[👈 Previous PR][prev]

[prev]: https://github.com/dfinity/ic/pull/4761