fix(rosetta): increase watchdog timeout (#4863)
The sync thread watchdog thread only waits for 10 seconds before
declaring the sync thread stale and restart it.

That timeout turns out to be too short for cases in which the IC is
going through instability, in which cases the block read operations can
take more than 10 seconds.

After looking at the latency metrics, it seems that 60 seconds provide a
good buffer for those cases. We're also making it configurable in case
such change needs to happen again without doing another release.