chore: Use same tone in contributing guides (#4865)
This PR is a follow up to previous changes that added the new
contributing guides for System API and Management Canister API
additions. The guides were mostly written in imperative tone but there
were some spots that used a different one. Consistently stick to more
imperative tone as the guides should have the steps that ought to be
performed for a proper addition to the System API/Management Canister
API.

Co-authored-by: Carly Gundy <47304080+cgundy@users.noreply.github.com>