When upgrading registry, decode using HighCapacityRegistryAtomicMutateRequest, rather than vanilla RegistryAtomicMutateRequest.
The main registry canister library compiles, but tests do NOT compile (yet).

These types are super similar. In fact, there are tests that show that they are highly compatible. That is, you can generally encode with one and decode with the other. The exception is, of course, when LargeValueChunkKeys is involved.