fix(ic-backup): [CON-1493] Only replay blocks until a CUP height during a backup (#4823)
Currently we don’t verify the state hash in a CUP if the CUP’s height is
below the latest state’s height:
[click](https://sourcegraph.com/github.com/dfinity/ic@23abac5891de0ebde5c49c0fe91a1aab39c6241f/-/blob/rs/replay/src/player.rs?L1197-1202).

When replaying blocks for the backup purposes, we start with a
checkpoint height `h`, we execute blocks until we reach the next CUP
height `h'`, and continue executing blocks until we reach the smallest
height `h'' >= h'` which is explicitly finalized:
[click](https://sourcegraph.com/github.com/dfinity/ic@23abac5891de0ebde5c49c0fe91a1aab39c6241f/-/blob/rs/replay/src/backup.rs?L460-468).
Only then we verify the CUP:
[click](https://sourcegraph.com/github.com/dfinity/ic@23abac5891de0ebde5c49c0fe91a1aab39c6241f/-/blob/rs/replay/src/player.rs?L1092-1104).

This means that if `h''` is strictly larger than `h'`, i.e., if the CUP
height was not explicitly finalized, we will not verify the state hash
in the CUP.

With this change we will stop executing blocks once we reach a CUP
height and as a consequence we will do the full CUP verification.

Note: this change only affects the code path in ic-replay which is
executed during backups.