feat: Make CanisterModule writable for snapshot upload (#4885)
This PR adds a `write` method to the `CanisterModule` type, such that
this binary data can be uploaded and written in slices via the snapshot
data upload endpoints.

Using the write method may make the `CanisterModule` invalid. But this
type seems to have no validity guarantees in existing code: E.g.,
[here](https://sourcegraph.com/github.com/dfinity/ic@mwe/module_write/-/blob/rs/execution_environment/src/canister_manager/types.rs?L213),
it's composed from user-provided wasm chunks. Also, the canister module
will only be compiled when put in the compilation cache, and only at
that point does the replica know if the module is valid.

When the user is done uploading slices, they may load the snapshot. At
this point, the snapshot's content, including the `CanisterModule`, must
be validated (in a follow-up PR).

The alternative considered was to change the `CanisterSnapshot` struct
into an enum which can represent both `Valid` and `Partial` canister
snapshots. However, the blast radius of this would be significant.