refactor: Move unflushed_changes out of snapshots (#4854)
CanisterSnapshots contains a field unflushed_changes that communicates
to the checkpoint logic what snapshot creations/deletions/restores
happened so that they can be correctly represented on disk.

The same field also needs to be used for some of the upcoming canister
migration logic, in particular when a canister is renamed locally.

It would be weird for the canister migration logic to populate a field
in CanisterSnapshots, so this refactor generalizes the concept of
"snapshot operations" to a more general concept of "checkpoint
operations", and moves the logic accordingly. There should be no change
in behaviour.

---------

Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>