test: add a stress and a soak test for the https outcalls feature (#4449)
This test adds an update message to the already existing
`proxy_canister`, which performs `count` concurrent outcalls.

However, as [the number of canister messages is limited to
500](https://internetcomputer.org/docs/building-apps/canister-management/resource-limits),
the requests have to be split into batches.

All requests are lightweight both in therms of request size and response
size.

The test sets up three testnets, with 13, 28 and 40 nodes respectively.
For each of them it tries to send 200, 500 and 1000 requests, and
measures the average qps for each (over 2 experiments).

Results are as follows:
```
QPS:
node# \ concurrency_level | 200 | 500 | 1000
13:                       | 52  |  85 | 81
28:                       | 26  |  53 | 55
40:                       | 51  |  67 | 65
```

The 13 and 28 nodes subnets were set up to better simulate production
subnets (package drops, rtts etc). The 28 node subnet saw a ~25%
decrease in qps when simulated under production conditions. Meaning we
can infer that a ~25% decrease in qps could be seen on the 40 nodes
subnet too, in a production environment, leading to ~40-45 qps.

Interpretation:
- The canister measures the time of the http_request to the management
canister. This includes ingress message processing, consensus, block
making etc etc.
- The 40 node subnet is slower mainly because consensus takes longer,
and there's a better chance of some adapters slightly slower
- Not fully saturating the 500 concurrency limit yields worse qps.
- In the current setup, the qps is also correlated with the delay of the
target server.

Colocating nodes doesn't seem to affect the average qps by a lot
(sometimes it's ~10% higher / lower, but there's no clear indication
that it's way better / worse), as the main bottleneck here is the 500
limit on the concurrent canister messages.

Future attempts at stressing the feature may include: 

1. Installing multiple proxy canisters on the same subnet to bypass the
500 limit
2. Fully saturating the bottleneck by continuously sending new requests
in the message queue instead of waiting for all 500 to return.
3. Sending requests to different servers, so that the warm up is no
longer helping much.