refactor(nns): Remove prost annotation on Governance API type (#4843)
# Why

In order to define API types not compatible with prost, we want to let
the Governance global state no longer need the prost derive.

# What

* Clean up all protobuf methods on NNS Governance, which are already
deprecated in different ways:
* `manage_neuron_pb`/`list_neurons_pb` have their panic probability
gradually ramped up to 100% for 2 weeks
* `transfer_notification_pb`/`claim_or_refresh_neuron_from_account_pb`
have been panicking for a long time
* Migrate some protobuf encoding/decoding tests from Governance to GTC
(since GTC still uses protobuf). Using any prost types can work, but
replacing with GTC seems to align with the original intention the best.
* Remove governance_mem_test since it has been "manual", and less useful
after neurons are migrated to stable memory (the purpose was to detect
bugs related to filling up the wasm memory).
* On `ic_nns_governance_api::pb::v1::Governance` derive, drop
`prost::Message` and `Comparable`, and add `Default` and `Debug` (as
they were provided by `prost::Message`). `Comparable` was only needed by
some tests against internal types.
* Delete `prost` macros on all the fields within Governance struct.
* Delete `ic-get-neuron-ids` as it's no longer possible to decode the
init state with protobuf. This is OK since it's not used within the
monorepo, and the binary is not exported outside of the monorepo.