feat(registry): During upgrade, decodes using HighCapacityRegistryAtomicMutateRequest. (#4794)
In a recent PR, this high-capacity type was used to ENCODE changelog
entries. This PR "completes that circle", so to speak.

# Secondary Change(s)

Doing this made me realize that I made a mistake in an earlier change.

Nothing that `from_serializable_form` calls (whether direct or indirect)
should chunkify. Whereas, earlier, I made `changelog_insert` do
chunkification (and `changelog_insert` is indirectly called by
`from_serialized_form`). This change fixes that by moving chunkification
"up" the call graph to `apply_mutations` (which is not called by
`from_serializable_form`, not even indirectly). This actually turns out
to be the right "choke point" where chunkification should be applied.

Fortunately, the new buggy `changelog_insert` behavior (i.e.
chunkification) was never enabled in release builds (feature flags
FTW!). Therefore, we need not issue any recalls.

# Background

There are a couple of reasons these things can be done in separate PRs:

1. The new way of encoding is behind a flag, and per our standard
practice, is disabled by default (in release builds). Later, we will
flip this flag so that this behavior is enabled in production.

2. Even if it were enabled, the result for all current data is
equivalent. This is because the new type is "highly compatible" with the
old type.

# References

Closes [NNS1-3767].

[NNS1-3767]: https://dfinity.atlassian.net/browse/NNS1-3767