test(Networking): Effective canister ID spec tests in rust (#4641)
This PR implements the effective canister id endpoint spec tests as rust
system tests.

The PR moves the `TestAgent` out of `http_endpoint/tests/common` into
it's own `test_agent` crate, such that it can be reused in other
settings.
It then adds a system test in `tests/networking`, that makes use of the
`test_agent` crate, to run the tests specified by the effective canister
id spec on a testnet with two single node subnets.