chore(node_rewards): migrate tests to pocket ic (#4938)
This migrates some tests from state machine to pocket IC, and adds
support for the nervous_system agent way of making requests so they can
be used as test helpers. Additionally, it adds another helper to get an
agent from StateMachine or PocketIc, to make constructing agents more
generic so that the test environment can be more easily switched in the
future.