test(nns): Within the golden state test, vote on proposals by impersonating well-known neurons (#4922)
# Why

The previous approach of minting u64::MAX/4 ICPs within the golden NNS
state tests will stop working if we are able to detect a voting power
spike and use previous ballots in such an event. Therefore, the only way
to pass for testing upgrades would be to vote with existing neurons.

# What

* Create versions of `nns_cast_vote` and `manage_neuron` without panics
* Create "super powerful neuron" with a smaller (but more realistic)
number of tokens
* Vote on upgrade proposals with well-known public neurons