refactor(nns): Remove all prost annotations on the NNS Governance API (#4906)
# Why

NNS Governance API does not use protobuf anymore (in a previous
[PR](https://github.com/dfinity/ic/pull/4843). Therefore, every
`prost:*` can be removed from the API types file.

# What

* Replace `prost::Message` with `Debug, Default`
* Replace `prost::OneOf` with `Debug`
* Remove `comparable::Comparable`, `compare_default` and
`self_describing` since they are needed for some tests which are based
on internal types.
* Remove all `#[prost(...)]` on fields
* Remove a helper function `get_all_proposal_ids` since it's not used
* Change from enum to i32 for some tests
* This might seem "inconvenient" or "less readable", but note that it
correctly reflects how the API is difficult to use right now, as the
meanings of those i32's are not part of the API
* `DisburseMaturityResponse::amount_disbursed_e8s` was incorrectly not
`pub` but using a method provided by prost. It is changed to `pub`
* Some `Option<T>` methods were accessed by methods using prost, which
automatically `unwrap_or_default`. Added unwrapping.
* Add `strum_macros::FromRepr` to enums when needed, and use
`E::from_repr` instead of `E::try_from` when the i32 forms are difficult
to use.