feat(nns): Validate the maturity disbursement index (#4924)
# Why

The maturity disbursement index is used to find the maturity
disbursement that are ready for finalization, as well as scheduling the
timer to run at the correct time. We want to make sure the index is
correct, by adding validation that the index is equivalent to the
primary data as much as possible.

# What

* Implement `MaturityDisbursementIndexValidator`
* Run the validator when `is_disburse_maturity_enabled`