test(sns): Check that NNS Governance and SNS-W can be upgraded without each other (#4950)
There was an instance where an NNS Governance upgrade without SNS-W
upgrade causing backward compatibility issue, where a field stops
getting populated within the CreateServiceNervousSystem payload, while
the corresponding change in SNS-W to validate this field is not
released. On a high level, there is some message passing between NNS
Governance and SNS-W for CSNS proposal execution, and we want to make
sure the message passing is not broken if either of them is upgraded
without each other.