feat(PSEC-2502): move dep-scan periodic job to different repo (#4990)
We want to move the periodic dependency scan job to a private repo
including the list of repositories which should be scanned. This allows
us to securely scan private DFINITY repos such as caffeine.

The idea is to leave all code in the `ic` repo and define a composite
action there, which can be invoked from another repo. The MR and release
scan jobs will continue to run in the `ic` repo.

This PR makes the required adjustments to the `ic` repo:

- disable "Dependency Scan Nightly" (will be removed later)
- define a composite action under
`ci/actions/dependencies/periodic/action.yml`
- remove the list of repositories which should be scanned