feat(IDX): add arm64-linux pocket-ic builds
This adds a CI build of pocket-ic-server on namespace runners. This is
inspired by the arm64-darwin builds we already run on namespace.

This is the first step is releasing pocket-ic artifacts on arm64-linux.
The artifact is currently built with debug assertions, and the artifact
is not uploaded to our CDN; these two issues will be addressed in
subsequent PRs.