feat(IDX): add arm64-linux pocket-ic builds (#4991)
This adds a CI build of pocket-ic-server on namespace runners. This is
inspired by the arm64-darwin builds we already run on namespace.

This is the first step in releasing pocket-ic artifacts on arm64-linux.
The artifact is currently built with debug assertions, and the artifact
is not uploaded to our CDN; these two issues will be addressed in
subsequent PRs.

NOTE: pocket-ic builds fine outside of the build container, so for
simplicity we run the build outside of the container.