feat(crypto): CRP-2777 enhance vetKD key lifecycle system test (#4901)
Enhances the vetKD system test so that many key properties of the vetKD
API are tested that were not tested before. For example:
* Two executions of the vetKD protocol must yield the same (decrypted)
vetKey for the same input/context/keyID/canister combination, even if
the key derivation happens on another subnet.
* Two executions of the vetKD protocol must yield different (decrypted)
vetKeys if any of input, context, keyID, canister are different.