chore(IDX): add os_info repository rule (#4994)
This adds a new repository rule that reads the OS info and exports it as
`os_info`. By doing this we can be more flexible in how we provide the
`remote_subdir` when uploading artifacts and are not limited to rules
only, but can also use a macro. This is part of the work for extracting
the upload step from the build itself.