feat(registry): Migrate `Registry::store` to high-capacity types. (#4913)
Also, make the `get_value` and `get_changes_since` canister methods use
high-capacity types to encode responses. As long as there is no
large/chunked records, the result is the same as before. And since
chunking is disabled, there is no externally visible difference in
behavior.

With this, the initial implementation of large records in the registry
canister is complete. There is still work to do before this can be
released, but this is nevertheless a significant mile stone! In
particular, clients still need to be upgraded to high-capacity types,
and this is already being worked on.

# Change Details

Added `pub fn get_high_capacity` to `Registry`. This is like
`Registry::get`, but can return chunked values.

The implementation of the `get` canister method relies on
`get_high_capacity`; whereas, before, it used to call the old
`Registry::get` method.

Since calling `Registry::get` no longer happens in `canister.rs`, I was
able to lower the visibility of `Registry::get` from `pub` to
`pub(crate)`.

`get` is now implemented in terms of `get_high_capacity`. As you can
imagine, the main thing that happens in the impl of `get` besides
calling `get_high_capacity` is dechunkification.

# References

Closes [NNS1-3683], and [NNS1-3701].

[NNS1-3683]: https://dfinity.atlassian.net/browse/NNS1-3683
[NNS1-3701]: https://dfinity.atlassian.net/browse/NNS1-3701